Fixed "get() on array" and "undefined method stdClass::get()"
I noticed that using `responseType` other than `collection` causes one of these errors 

For responseType `array` Error: `get() on array`
and for `object` Error: `undefined method stdClass::get()`

As it was hardcoded on the `getToken` method to use laravel's `get` helper, which is a collection helper and doesn't work with `Arrays` or `Objects`. 
So using responseType other than Collection was causing this error.